# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Hafidz App BFF API
  version: 1.0.0
  description: Backend-for-Frontend + AI Voice Ayah Finder for Hafidz App. See docs/05-BACKEND-API-SPEC.md.
servers:
  - url: http://localhost:8000
    description: dev
  - url: https://api.example.com
    description: prod
components:
  parameters:
    DeviceId:
      name: X-Device-Id
      in: header
      required: true
      schema: { type: string, format: uuid }
...
  schemas:
    Problem:
      type: object
      required: [type, title, status, code]
      properties:
        type: { type: string }
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
        code:
          type: string
          enum: [AUDIO_TOO_SHORT, AUDIO_TOO_LONG, AUDIO_UNSUPPORTED, AUDIO_SILENT, NOT_FOUND, UPSTREAM_UNAVAILABLE, RATE_LIMITED, VALIDATION_ERROR]
        request_id: { type: string }
    MatchSpan:
      type: object
      required: [ayah, word_start, word_end]
      properties:
        ayah: { type: integer }
        word_start: { type: integer }
        word_end: { type: integer }
    AyahMatch:
      type: object
      required: [key, surah, ayah_start, ayah_end, page, juz, score]
      properties:
        key: { type: string, example: '2:255' }
        surah: { type: integer, minimum: 1, maximum: 114 }
        ayah_start: { type: integer, minimum: 1 }
        ayah_end: { type: integer, minimum: 1 }
        page: { type: integer, minimum: 1, maximum: 604 }
        juz: { type: integer, minimum: 1, maximum: 30 }
        surah_name_arabic: { type: string }
        surah_name_latin: { type: string }
        surah_name_translation: { type: [string, 'null'] }
        score: { type: number, minimum: 0, maximum: 1 }
        match_span:
          oneOf:
            - $ref: '#/components/schemas/MatchSpan'
            - type: 'null'
    VoiceDetectResponse:
      type: object
      required: [request_id, transcript, confidence, margin, ambiguous, reason, best, candidates, thresholds, timing_ms]
      properties:
        request_id: { type: string }
        transcript: { type: string }
        confidence: { type: number, minimum: 0, maximum: 1 }
        margin: { type: number }
        ambiguous: { type: boolean }
        reason:
          type: [string, 'null']
          enum: [too_short, basmala_only, identical_ayat, low_confidence, null]
        best:
          oneOf:
            - $ref: '#/components/schemas/AyahMatch'
            - type: 'null'
        candidates:
          type: array
          items: { $ref: '#/components/schemas/AyahMatch' }
          maxItems: 5
        thresholds:
          type: object
          properties:
            auto: { type: number, example: 0.8 }
            min: { type: number, example: 0.55 }
            margin: { type: number, example: 0.1 }
        timing_ms:
          type: object
          properties:
            decode: { type: integer }
            asr: { type: integer }
            match: { type: integer }
            total: { type: integer }
...
paths:
  /v1/voice/detect:
    post:
      tags: [voice]
      summary: Detect Qur'an ayah(s) from a short recitation audio clip
      parameters: [ { $ref: '#/components/parameters/DeviceId' } ]
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [audio]
              properties:
                audio: { type: string, format: binary }
                hint_surah: { type: integer, minimum: 1, maximum: 114 }
                locale: { type: string, enum: [id, en], default: id }
                max_candidates: { type: integer, minimum: 1, maximum: 5, default: 5 }
      responses:
        '200':
          description: Detection result
          content:
            application/json:
              schema: { $ref: '#/components/schemas/VoiceDetectResponse' }
        '400': { description: Invalid form values, content: { application/problem+json: { schema: { $ref: '#/components/schemas/Problem' } } } }
        '413': { description: Too large/long, content: { application/problem+json: { schema: { $ref: '#/components/schemas/Problem' } } } }
        '415': { description: Unsupported audio, content: { application/problem+json: { schema: { $ref: '#/components/schemas/Problem' } } } }
        '422': { description: Too short / silent, content: { application/problem+json: { schema: { $ref: '#/components/schemas/Problem' } } } }
        '429': { description: Rate limited, content: { application/problem+json: { schema: { $ref: '#/components/schemas/Problem' } } } }
        '503': { description: Rate limiter or detection data unavailable, content: { application/problem+json: { schema: { $ref: '#/components/schemas/Problem' } } } }
...